lambda/requirements.txt
boto3==1.36.25
pandas==2.2.3
snowflake-connector-python==3.13.2